Add swipe to refresh to shopping lists
