+++ features/shopping_lists/build.gradle.kts
@@ -24,17 +24,18 @@ dependencies {
     implementation(project(":database"))
 
     implementation(libs.android.material.material)
+    implementation(libs.androidx.compose.material)
 
-    implementation(platform(libs.android.compose.bom))
-    implementation(libs.android.compose.material3)
-    implementation(libs.android.compose.ui.toolingPreview)
-    implementation(libs.android.compose.runtime.livedata)
+    implementation(platform(libs.androidx.compose.bom))
+    implementation(libs.androidx.compose.material3)
+    implementation(libs.androidx.compose.ui.toolingPreview)
+    implementation(libs.androidx.compose.runtime.livedata)
     implementation(libs.androidx.lifecycle.viewmodelCompose)
     implementation(libs.google.accompanist.themeadapter.material3)
-    debugImplementation(libs.android.compose.ui.tooling)
-    debugImplementation(libs.android.compose.ui.testManifest)
-    androidTestImplementation(libs.android.compose.ui.testJunit)
-    androidTestImplementation(platform(libs.android.compose.bom))
+    debugImplementation(libs.androidx.compose.ui.tooling)
+    debugImplementation(libs.androidx.compose.ui.testManifest)
+    androidTestImplementation(libs.androidx.compose.ui.testJunit)
+    androidTestImplementation(platform(libs.androidx.compose.bom))
 
     implementation(libs.google.dagger.hiltAndroid)
     kapt(libs.google.dagger.hiltCompiler)
