+++ features/shopping_lists/src/main/kotlin/gq/kirmanak/mealient/shopping_lists/network/ShoppingListsDataSource.kt
@@ -1,5 +1,7 @@
 package gq.kirmanak.mealient.shopping_lists.network
 
+import gq.kirmanak.mealient.shopping_lists.models.ShoppingListsInfo
+
 interface ShoppingListsDataSource {
 
     suspend fun getPage(page: Int, perPage: Int): ShoppingListsInfo
