+++ gradle/libs.versions.toml
@@ -91,19 +91,23 @@ composeKotlinCompilerExtension = "1.3.2"
 accompanistVersion = "0.28.0"
 # https://mvnrepository.com/artifact/androidx.paging/paging-compose
 pagingCompose = "1.0.0-alpha17"
+# https://developer.android.com/jetpack/androidx/releases/compose-material
+materialCompose = "1.3.1"
 
 [libraries]
 android-gradlePlugin = { group = "com.android.tools.build", name = "gradle", version.ref = "androidGradlePlugin" }
 android-tools-desugar = { group = "com.android.tools", name = "desugar_jdk_libs", version.ref = "desugar" }
 android-material-material = { group = "com.google.android.material", name = "material", version.ref = "material" }
 
-android-compose-bom = { group = "androidx.compose", name = "compose-bom", version.ref = "composeBom" }
-android-compose-material3 = { group = "androidx.compose.material3", name = "material3" }
-android-compose-ui-toolingPreview = { group = "androidx.compose.ui", name = "ui-tooling-preview" }
-android-compose-ui-tooling = { group = "androidx.compose.ui", name = "ui-tooling" }
-android-compose-ui-testJunit = { group = "androidx.compose.ui", name = "ui-test-junit4" }
-android-compose-ui-testManifest = { group = "androidx.compose.ui", name = "ui-test-manifest" }
-android-compose-runtime-livedata = { group = "androidx.compose.runtime", name = "runtime-livedata" }
+androidx-compose-bom = { group = "androidx.compose", name = "compose-bom", version.ref = "composeBom" }
+androidx-compose-material3 = { group = "androidx.compose.material3", name = "material3" }
+androidx-compose-ui-toolingPreview = { group = "androidx.compose.ui", name = "ui-tooling-preview" }
+androidx-compose-ui-tooling = { group = "androidx.compose.ui", name = "ui-tooling" }
+androidx-compose-ui-testJunit = { group = "androidx.compose.ui", name = "ui-test-junit4" }
+androidx-compose-ui-testManifest = { group = "androidx.compose.ui", name = "ui-test-manifest" }
+androidx-compose-runtime-livedata = { group = "androidx.compose.runtime", name = "runtime-livedata" }
+
+androidx-compose-material = { group = "androidx.compose.material", name = "material", version.ref = "materialCompose" }
 
 google-accompanist-themeadapter-material3 = { group = "com.google.accompanist", name = "accompanist-themeadapter-material3", version.ref = "accompanistVersion" }
